DashBoard 65% refatorada.
